pyproject.toml
[project]
name = "efficient-cv"
version = "0.0.1"
description = "A library for efficient computer vision on edge devices."
authors = [
    { name = "Michael Krah", email = "mickra@bu.edu" },
    { name = "Zach Gentile", email = "zgentile@bu.edu" },
    { name = "Alex Lavaee", email = "alavaee@bu.edu" },
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "backoff==2.2.1",
    "botocore==1.35.77",
    "certifi==2024.8.30",
    "charset-normalizer==3.4.0",
    "colorama==0.4.6",
    "deprecation==2.1.0",
    "filelock==3.13.1",
    "fsspec==2024.2.0",
    "h5py==3.12.1",
    "idna==3.10",
    "Jinja2==3.1.3",
    "jmespath==1.0.1",
    "MarkupSafe==2.1.5",
    "matplotlib>=3.10.1",
    "mpmath==1.3.0",
    "networkx==3.2.1",
    "numpy==1.26.3",
    "packaging==24.2",
    "pandas>=2.2.3",
    "pillow==10.2.0",
    "prettytable==3.12.0",
    "protobuf==3.20.3",
    "python-dateutil==2.9.0.post0",
    "qai-hub==0.20.0",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "s3transfer==0.10.4",
    "scipy>=1.15.2",
    "semver==3.0.2",
    "six==1.17.0",
    "sympy==1.13.1",
    "timm>=1.0.15",
    "torch==2.5.1",
    "torchaudio==2.5.1",
    "torchvision==0.20.1",
    "tqdm==4.67.1",
    "typing_extensions==4.12.2",
    "urllib3==2.2.3",
    "wcwidth==0.2.13",
    "omegaconf",
    "transformers",
    "accelerate",
    "open_clip_torch",
    "einops",
    "scipy",
    "pillow",
    "gdown",
    "huggingface-hub",
    "einops",
    "wandb",
    "torch-fidelity",
    "torchinfo",
    "webdataset",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu", marker = "sys_platform != 'linux'" },
    { index = "pytorch-cu124", marker = "sys_platform == 'linux'" },
]
torchvision = [
    { index = "pytorch-cpu", marker = "sys_platform != 'linux'" },
    { index = "pytorch-cu124", marker = "sys_platform == 'linux'" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
